+++ crates/node/src/constants.rs
@@ -1,2 +1,6 @@
 pub const DEFAULT_CONFIGS_URL: &str =
     "https://raw.githubusercontent.com/keep-starknet-strange/madara/main/configs/index.json";
+
+pub const DEV_CHAIN_ID: &str = "dev";
+pub const SHARINGAN_CHAIN_ID: &str = "sharingan";
+pub const LOCAL_TESTNET_CHAIN_ID: &str = "local_testnet";